+++ python/lvmecp/actor/commands/engineering.py
@@ -35,12 +35,22 @@ async def get_eng_mode_status(actor: ECPActor) -> dict:
     else:
         ends_at = started_at + duration
 
+    sw_bypass = registers["bypass_software_status"]
+    sw_is_remote = sw_bypass and registers["bypass_software_remote"]
+    sw_bypass_mode = "none" if not sw_bypass else "remote" if sw_is_remote else "local"
+
+    hw_bypass = registers["bypass_hardware_status"]
+    hw_is_remote = hw_bypass and registers["bypass_hardware_remote"]
+    hw_bypass_mode = "none" if not hw_bypass else "remote" if hw_is_remote else "local"
+
     return {
         "enabled": enabled,
         "started_at": timestamp_to_iso(started_at),
         "ends_at": timestamp_to_iso(ends_at),
-        "plc_software_override": registers["engineering_mode_software_status"],
-        "plc_hardware_override": registers["engineering_mode_hardware_status"],
+        "plc_software_bypass": sw_bypass,
+        "plc_hardware_bypass": hw_bypass,
+        "plc_software_bypass_mode": sw_bypass_mode,
+        "plc_hardware_bypass_mode": hw_bypass_mode,
     }
 
 
@@ -60,20 +70,20 @@ def engineering_mode():
     "If not passed, the default timeout is used.",
 )
 @click.option(
-    "--hardware-override",
+    "--hardware-bypass",
     is_flag=True,
-    help="Sets the hardware override flag.",
+    help="Sets the hardware bypass flag.",
 )
 @click.option(
-    "--software-override",
+    "--software-bypass",
     is_flag=True,
-    help="Sets the software override flag.",
+    help="Sets the software bypass flag.",
 )
 async def enable(
     command: ECPCommand,
     timeout: float | None = None,
-    hardware_override: bool = False,
-    software_override: bool = False,
+    hardware_bypass: bool = False,
+    software_bypass: bool = False,
 ):
     """Enables the engineering mode."""
 
@@ -82,10 +92,10 @@ async def enable(
     await command.actor.engineering_mode(True, timeout=timeout)
     await asyncio.sleep(0.5)  # Allow time for the e-mode task to run.
 
-    if hardware_override:
-        await modbus.write_register("engineering_mode_hardware_remote", True)
-    if software_override:
-        await modbus.write_register("engineering_mode_software_remote", True)
+    if hardware_bypass:
+        await modbus.write_register("bypass_hardware_remote", True)
+    if software_bypass:
+        await modbus.write_register("bypass_software_remote", True)
 
     return command.finish(engineering_mode=await get_eng_mode_status(command.actor))
 
@@ -98,8 +108,8 @@ async def disable(command: ECPCommand):
 
     await command.actor.engineering_mode(False)
 
-    await modbus.write_register("engineering_mode_hardware_remote", False)
-    await modbus.write_register("engineering_mode_software_remote", False)
+    await modbus.write_register("bypass_hardware_remote", False)
+    await modbus.write_register("bypass_software_remote", False)
 
     return command.finish(engineering_mode=await get_eng_mode_status(command.actor))